+++ src/ts_platform/models/__init__.py
@@ -2,10 +2,12 @@
 
 from ts_platform.models.base import BaseForecastModel
 from ts_platform.models.linear import LinearForecastModel
+from ts_platform.models.linear_family import DLinearForecastModel, NLinearForecastModel
 from ts_platform.models.mlp import MLPForecastModel
 from ts_platform.models.moving_average import MovingAverageForecastModel
 from ts_platform.models.naive import NaiveLastValueModel
 from ts_platform.models.nbeats import NBeatsForecastModel
+from ts_platform.models.patchtst import PatchTSTForecastModel
 from ts_platform.models.recurrent import GRUForecastModel, LSTMForecastModel, RNNForecastModel
 from ts_platform.models.registry import MODEL_REGISTRY, build_model
 from ts_platform.models.seasonal_naive import SeasonalNaiveForecastModel
@@ -15,6 +17,7 @@ from ts_platform.models.transformer import TransformerForecastModel
 __all__ = [
     "MODEL_REGISTRY",
     "BaseForecastModel",
+    "DLinearForecastModel",
     "GRUForecastModel",
     "LSTMForecastModel",
     "LinearForecastModel",
@@ -22,6 +25,8 @@ __all__ = [
     "MovingAverageForecastModel",
     "NBeatsForecastModel",
     "NaiveLastValueModel",
+    "NLinearForecastModel",
+    "PatchTSTForecastModel",
     "RNNForecastModel",
     "SeasonalNaiveForecastModel",
     "TCNForecastModel",